Feedback: Add types file, create init function in MetricConfig component
