+++ publisher/src/components/MetricConfiguration/types.ts
@@ -15,10 +15,29 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-export * from "./Configuration";
-export * from "./ContextConfiguration";
-export * from "./MetricBox";
-export * from "./MetricConfiguration";
-export * from "./MetricConfiguration.styles";
-export * from "./MetricDefinitions";
-export * from "./types";
+import {
+  MetricConfigurationSettingsOptions,
+  MetricContext,
+} from "@justice-counts/common/types";
+
+export type MetricSettings = {
+  key: string;
+  enabled?: boolean;
+  settings?: { key: string; included: MetricConfigurationSettingsOptions }[];
+  contexts?: {
+    key: string;
+    value: MetricContext["value"];
+  }[];
+  disaggregations?: {
+    key: string;
+    enabled?: boolean;
+    dimensions?: {
+      key: string;
+      enabled?: boolean;
+      settings?: {
+        key: string;
+        included: MetricConfigurationSettingsOptions;
+      }[];
+    }[];
+  }[];
+};
